Add: Rename project, refactor code and reach quite stable state
A bunch of features have been added and the code have been almost
entirely refactored.

Here is a non exhaustive list of new features:

- Completion (triggered with `TAB`) based on non-regex pattern. If several candidates are found, `skim` is used to filter them.
- Basic alias support including global alias: Alias can be anywhere in the command line
- Completion hints based on history (like fish or zsh-auto-suggestions)
- Error code now available thanks to the `status` builtin command
- Prompt color changes based on error code
- Log feature is now functional
- Some additional keybindings have been implemented, such as:
	- `CTRL + f`: Accept completion hint
	- `CTRL + o`: Enter
	- `ALT + f`: Go forward a word
	- `ALT + w`: Go backward a word
	- `ALT + u`: Undo (That is soooooo cool)

There is still a lot of work to do regarding splitting code across
several files, but it will be performed in later commits.

The project has been renamed to OSH since it has now reached a state
where it could be used as a shell for almost all every day cases.